fix(#1453): fmt value.rs + campsite-split node value tests into sibling file
Full gate flagged two failures in bindings/node/src/value.rs:
- fmt: cargo fmt --all resolves the formatting diff.
- file-size: value.rs had grown to 812 lines (>800 campsite limit).
  Move the inline #[cfg(test)] mod tests block out to a sibling
  bindings/node/src/value_tests.rs, wired via
  #[cfg(test)] #[path = "value_tests.rs"] mod tests; — value.rs is now
  642 lines. Same 12 unit tests run (value::tests::*), unchanged coverage.
